Removed automatic translation. Added component bundling and mount functionality.
Babel translation is now opt-in, rather than the default. Rather than using Babel's require hook to load components, it is now used as part of the bundling process. This removes the massive overhead that it introduced when loading files in the same process as the renderer. It also removes the requirement for translated files to use the `jsx` extension, which had been used as a whitelist hack to get around the aforementioned overhead.

If you want to continue as before, just add `translate=True` to your render_component calls.

The bundling/mounting functionality is similar to what previously existed when django-react and django-webpack were more tightly coupled. The primary difference is that it is now opt-in, rather than the default.

Re #24